+++ Canvas_Game/src/LoginPage.jsx
@@ -2,8 +2,9 @@ import React, { useState } from 'react';
 import styles from './LoginPage.module.css';
 import InputField from './InputField';
 import wizardLogo from './assets/WizardLogo.png';
+import { Dashboard } from './scenes';
 
-const LoginPage = ({ switchToSignUp }) => { 
+const LoginPage = ({switchToSignUp,switchToDashboard}) => { 
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
 
@@ -36,10 +37,8 @@ const LoginPage = ({ switchToSignUp }) => {
           onChange={handlePasswordChange}
         />
         <div className={styles.buttonGroup}>
-          <button type="submit" className={styles.button}>Sign In</button>
-          <button type="button" onClick={switchToSignUp} className={styles.signUpButton}>
-            Sign Up
-          </button>
+          <button type="submit" onClick={switchToDashboard}className={styles.signUpButton}>Sign In </button>
+          <button type="button" onClick={switchToSignUp} className={styles.signUpButton}>Sign Up </button>
         </div>
       </form>
     </main>
